+++ frontend/src/utility_components/ConfirmModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../Button';
+import Button from './Button';
 
 type ConfirmModalProps = {
     header: string;
